+++ common/kernels/calcSobel_dY_kernel.cu
@@ -41,7 +41,7 @@ __global__ void calcSobel_dY_k1(float* inputPixels, float* intermediate,
     }
     float kernelSum_1 = 4.0;
     float kernelSum_2 = 2.0;
-    extern __shared__ float ghb[]; // for per-thread local history
+    extern __shared__ int ghb[]; // for per-thread local history
     int my_ghb_index = ((threadIdx.y * blockDim.x) + threadIdx.x) * 3;
 
     // still check for this in case of small img, not all threads need execute
@@ -71,9 +71,9 @@ __global__ void calcSobel_dY_k1(float* inputPixels, float* intermediate,
                     // finish up last few values with NUM_TEX reads
                     for(int ii = (RADIUS-NUM_TEX_SCALED+1); ii <= RADIUS;ii++) {
                         int filterWeightLoc = RADIUS + ii;
-                        float curValueHash = hashGHB(&ghb[my_ghb_index]);
-                        float texVal = tex1D<float>(tref,curValueHash);
-                        tmp += (ghb[my_ghb_index+2] + texVal) * kernel_2[filterWeightLoc];
+                        int curValueHash = hashGHB(&ghb[my_ghb_index]);
+                        int texVal = tex1D<int>(tref,curValueHash);
+                        tmp += (float)(ghb[my_ghb_index+2] + texVal) * kernel_2[filterWeightLoc];
                     }
                     float avg = (float)tmp / kernelSum_2;
                     intermediate[curElement] = avg;
@@ -115,7 +115,7 @@ __global__ void calcSobel_dY_k2(float* intermediate, float* outputPixels,
     }
     float kernelSum_1 = 4;
     float kernelSum_2 = 2;
-    extern __shared__ float ghb[]; // for per-thread local history
+    extern __shared__ int ghb[]; // for per-thread local history
     int my_ghb_index = ((threadIdx.y * blockDim.x) + threadIdx.x) * 3;
 
     // still check for this in case of small img, not all threads need execute
@@ -145,9 +145,9 @@ __global__ void calcSobel_dY_k2(float* intermediate, float* outputPixels,
                     // finish up last few values with NUM_TEX reads
                     for(int ii = (RADIUS-NUM_TEX_SCALED+1); ii <= RADIUS;ii++) {
                         int filterWeightLoc = RADIUS + ii;
-                        float curValueHash = hashGHB(&ghb[my_ghb_index]);
-                        float texVal = tex1D<float>(tref,curValueHash);
-                        tmp += (ghb[my_ghb_index+2] + texVal) * kernel_1[filterWeightLoc];
+                        int curValueHash = hashGHB(&ghb[my_ghb_index]);
+                        int texVal = tex1D<int>(tref,curValueHash);
+                        tmp += (float)(ghb[my_ghb_index+2] + texVal) * kernel_1[filterWeightLoc];
                     }
                     float avg = (float)tmp / kernelSum_1;
                     outputPixels[curElement] = avg;
